Move api util to root level test/utils
