+++ framework/test/mocha/functional/http/get/blocks.js
@@ -19,7 +19,7 @@ require('../../functional');
 const accountFixtures = require('../../../fixtures/accounts');
 const waitFor = require('../../../../utils/wait_for');
 const SwaggerEndpoint = require('../../../common/swagger_spec');
-const apiHelpers = require('../../../common/helpers/api');
+const apiHelpers = require('../../../../utils/api');
 const { Slots } = require('../../../../../src/modules/chain/dpos');
 
 const expectSwaggerParamError = apiHelpers.expectSwaggerParamError;
