+++ framework/test/mocha/functional/http/get/delegates.js
@@ -27,7 +27,7 @@ const accountFixtures = require('../../../fixtures/accounts');
 const randomUtil = require('../../../../utils/random');
 const waitFor = require('../../../../utils/wait_for');
 const SwaggerEndpoint = require('../../../common/swagger_spec');
-const apiHelpers = require('../../../common/helpers/api');
+const apiHelpers = require('../../../../utils/api');
 const { Slots } = require('../../../../../src/modules/chain/dpos');
 const { getNetworkIdentifier } = require('../../../common/network_identifier');
