+++ framework/test/mocha/functional/http/get/error_404.js
@@ -15,7 +15,7 @@
 'use strict';
 
 require('../../functional');
-const apiHelpers = require('../../../common/helpers/api');
+const apiHelpers = require('../../../../utils/api');
 
 describe('GET /unknown_endpoint', () => {
 	it('should fail with error 404', async () => {
