+++ framework/test/mocha/functional/http/get/node/transactions_pending.js
@@ -21,7 +21,7 @@ const {
 	registerSecondPassphrase,
 	registerMultisignature,
 } = require('@liskhq/lisk-transactions');
-const apiHelpers = require('../../../../common/helpers/api');
+const apiHelpers = require('../../../../../utils/api');
 const randomUtil = require('../../../../../utils/random');
 const SwaggerEndpoint = require('../../../../common/swagger_spec');
 const accountFixtures = require('../../../../fixtures/accounts');
