+++ framework/test/mocha/functional/http/get/peers.js
@@ -18,7 +18,7 @@ require('../../functional');
 const { P2P } = require('@liskhq/lisk-p2p');
 const { generatePeerHeader } = require('../../../common/generatePeerHeader');
 const SwaggerEndpoint = require('../../../common/swagger_spec');
-const apiHelpers = require('../../../common/helpers/api');
+const apiHelpers = require('../../../../utils/api');
 
 const expectSwaggerParamError = apiHelpers.expectSwaggerParamError;
