+++ framework/test/mocha/functional/http/get/transactions.js
@@ -25,7 +25,7 @@ const { getAddressFromPublicKey } = require('@liskhq/lisk-cryptography');
 const accountFixtures = require('../../../fixtures/accounts');
 const randomUtil = require('../../../../utils/random');
 const waitFor = require('../../../../utils/wait_for');
-const apiHelpers = require('../../../common/helpers/api');
+const apiHelpers = require('../../../../utils/api');
 const SwaggerEndpoint = require('../../../common/swagger_spec');
 const { Slots } = require('../../../../../src/modules/chain/dpos');
 const Scenarios = require('../../../common/scenarios');
