+++ framework/test/mocha/functional/http/post/1.second_secret.js
@@ -22,7 +22,7 @@ const {
 const Promise = require('bluebird');
 const phases = require('../../../common/phases');
 const accountFixtures = require('../../../fixtures/accounts');
-const apiHelpers = require('../../../common/helpers/api');
+const apiHelpers = require('../../../../utils/api');
 const randomUtil = require('../../../../utils/random');
 const waitFor = require('../../../../utils/wait_for');
 const apiCodes = require('../../../../../src/modules/http_api/api_codes');
