+++ framework/test/mocha/functional/http/post/7.dapps.out_transfer.js
@@ -23,7 +23,7 @@ const phases = require('../../../common/phases');
 const randomUtil = require('../../../../utils/random');
 const waitFor = require('../../../../utils/wait_for');
 const elements = require('../../../../utils/elements');
-const apiHelpers = require('../../../common/helpers/api');
+const apiHelpers = require('../../../../utils/api');
 const apiCodes = require('../../../../../src/modules/http_api/api_codes');
 const common = require('./common');
 const { getNetworkIdentifier } = require('../../../common/network_identifier');
