+++ framework/test/mocha/functional/http/put/node.js
@@ -17,7 +17,7 @@
 require('../../functional');
 const genesisDelegates = require('../../../data/genesis_delegates.json');
 const SwaggerEndpoint = require('../../../common/swagger_spec');
-const apiHelpers = require('../../../common/helpers/api');
+const apiHelpers = require('../../../../utils/api');
 
 const expectSwaggerParamError = apiHelpers.expectSwaggerParamError;
