+++ framework/test/utils/wait_for.js
@@ -17,7 +17,7 @@
 const popsicle = require('popsicle');
 const async = require('async');
 const Promise = require('bluebird');
-const apiHelpers = require('../mocha/common/helpers/api');
+const apiHelpers = require('../utils/api');
 const { Slots } = require('../../src/modules/chain/dpos');
 
 const { ACTIVE_DELEGATES } = global.constants;
